Gets rid of warnings about conversions to float.

+++ src/viewport.cpp
@@ -11,7 +11,7 @@ void Badger::Viewport::draw(Renderer* renderer) {
       Tile* tile = _map->tile(x, y);
       double coords[4];
       _map->spriteSheet()->textureCoordinates(tile->spriteIndex, coords);
-      renderer->drawSquare(x*32, y*32, 32, 32, coords[0], coords[1], coords[2], coords[3]);
+      renderer->drawSquare((float)x*32, (float)y*32, (float)32, (float)32, coords[0], coords[1], coords[2], coords[3]);
     }
   }
 }
